ping pong

+++ system_programing/ping_pong/ex1.c
@@ -5,16 +5,18 @@
  *  Approved By:  ;                                             *	 
  *  Description: signals Ping Pong                              */
 #define _XOPEN_SOURCE
-#include <stdlib.h> /* EXIT_FAILURE */
+#include <stdlib.h>    /* EXIT_FAILURE */
 #include <sys/types.h> /* pid_t */
-#include <unistd.h> /* pause */
-#include <signal.h> /* struct sigaction  */ 
-#include <stdio.h> /* printf */ 
+#include <unistd.h>    /* pause */
+#include <signal.h>    /* struct sigaction  */
+#include <stdio.h>     /* printf */
 /*------------------------------ exercise func --------------------------------*/
 
 static void PingPong();
 void Handler(int);
 /*------------------------------ implementetion --------------------------------*/
+
+volatile int flag = 1;
 
 int main()
 {
@@ -27,47 +29,51 @@ static void PingPong()
 {
     pid_t child_pid = {0};
     struct sigaction handler = {0};
-
+    int status = 0;
     handler.sa_handler = Handler;
 
     sigaction(SIGUSR1, &handler, NULL);
     sigaction(SIGUSR2, &handler, NULL);
 
-
     child_pid = fork();
 
-    if (child_pid == -1)
+    while (!status)
     {
-        exit(EXIT_FAILURE);
-    }
-    else if (child_pid > 0)
-    {
-        while (1)
+        if (child_pid == -1)
         {
-            printf("PING\n");
-            pause();
-            sleep(1);
-            kill(child_pid, SIGUSR1);
+            exit(EXIT_FAILURE);
+        }
+        else if (child_pid > 0)
+        {
+            if (flag == 1)
+            {
+                flag = 0;
+                status = kill(child_pid, SIGUSR1);
+                printf("\033[0;33m");
+                printf("PING\n");
+                printf("\033[0m");
+            }
+        }
+        else
+        {
+            if (flag == 1)
+            {
+                flag = 0;
+                status = kill(getppid(), SIGUSR2);
+                printf("\033[0;35m");
+                printf("PONG\n");
+                printf("\033[0m");
+            }
         }
     }
-    else
-    {
-
-        while (1)
-        {
-            printf("PONG\n");
-            kill(getppid(), SIGUSR2);
-            pause();
-            sleep(1);
-
-        }
-    }
-
     return;
 }
 
 void Handler(int num)
 {
     (void)num;
+
+    flag = 1;
+
     return;
 }
